redesign flash screen according to  mvp structure
