+++ app/src/main/AndroidManifest.xml
@@ -36,7 +36,7 @@ limitations under the License.
                 android:label="@string/app_name">
         </activity>
         <activity
-                android:name=".activities.FlashScreenActivity"
+                android:name=".ui.FlashScreenActivity"
                 android:label="@string/title_activity_flash_screen">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
